# specs/004-institution-iam-auth-integration/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Clinica Digital IAM API
  version: 0.3.0
  description: Contratos de IAM nativo in-app, registro institucional, sessao opaca, contexto do Shell e validacao RNDS para Organization, Location, Practitioner e PractitionerRole.
servers:
  - url: http://localhost:8080
paths:
...
  /api/admin/audit:
    get:
      summary: Lista trilha de auditoria por tenant
      operationId: listAuditEvents
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Eventos de auditoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEventSummary'
...
components:
  schemas:
...
    AuditEventSummary:
      type: object
      required: [id, eventType, outcome]
      properties:
        id:
          type: string
          format: uuid
        actorUserId:
          type: string
          format: uuid
          nullable: true
        actorPractitionerId:
          type: string
          format: uuid
          nullable: true
        eventType:
          type: string
        outcome:
          type: string
        traceId:
          type: string
          nullable: true
        payload:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
...